+++ src/XrmMockupShared/XrmMockupBase.cs
@@ -8,6 +8,8 @@ using Microsoft.Xrm.Sdk.Metadata;
 using Microsoft.Xrm.Sdk.Query;
 using Microsoft.Crm.Sdk.Messages;
 using System.IO;
+using System.Diagnostics;
+using System.Collections.ObjectModel;
 
 namespace DG.Tools.XrmMockup
 {
@@ -50,28 +52,50 @@ namespace DG.Tools.XrmMockup
         private Core Core;
         private MockupServiceProviderAndFactory ServiceFactory;
 
+        private readonly Dictionary<string, long> timers;
+        public IReadOnlyDictionary<string, long> Timers => new ReadOnlyDictionary<string, long>(timers);
+
+        protected XrmMockupSettings Settings { get; }
+        protected MetadataSkeleton Metadata { get; }
+        protected List<Entity> Workflows { get; }
+        protected List<SecurityRole> SecurityRoles { get; }
 
         /// <summary>
         /// Create a new XrmMockup instance
         /// </summary>
-        /// <param name="settings"></param>
-        protected XrmMockupBase(XrmMockupSettings settings) {
+        protected XrmMockupBase(XrmMockupSettings settings, MetadataSkeleton metadata = null, List<Entity> workflows = null, List<SecurityRole> securityRoles = null) {
+            timers = new Dictionary<string, long>();
+            Settings = settings;
 
-            var metadataDirectory = "../../Metadata/";
-            if (settings.MetadataDirectoryPath != null)
-                metadataDirectory = settings.MetadataDirectoryPath;
-            MetadataSkeleton metadata = Utility.GetMetadata(metadataDirectory);
+            Stopwatch stopwatch = Stopwatch.StartNew();
+            var metadataDirectory = settings.MetadataDirectoryPath ?? "../../Metadata/";
+            Metadata = metadata ?? Utility.GetMetadata(metadataDirectory);
+            timers[nameof(Utility.GetMetadata)] = stopwatch.ElapsedMilliseconds;
 
-            List<Entity> workflows = Utility.GetWorkflows(metadataDirectory);
-            List<SecurityRole> securityRoles = Utility.GetSecurityRoles(metadataDirectory);
+            stopwatch.Restart();
+            Workflows = workflows ?? Utility.GetWorkflows(metadataDirectory);
+            timers[nameof(Utility.GetWorkflows)] = stopwatch.ElapsedMilliseconds;
 
-            this.Core = new Core(settings, metadata, workflows, securityRoles);
-            this.ServiceFactory = new MockupServiceProviderAndFactory(this.Core);
+            stopwatch.Restart();
+            SecurityRoles = securityRoles ?? Utility.GetSecurityRoles(metadataDirectory);
+            timers[nameof(Utility.GetSecurityRoles)] = stopwatch.ElapsedMilliseconds;
+
+            stopwatch.Restart();
+            Core = new Core(settings, Metadata, Workflows, SecurityRoles);
+            timers[nameof(Core)] = stopwatch.ElapsedMilliseconds;
+
+            stopwatch.Restart();
+            ServiceFactory = new MockupServiceProviderAndFactory(Core);
+            timers[nameof(ServiceFactory)] = stopwatch.ElapsedMilliseconds;
+
             if (settings.EnableProxyTypes == true) {
+                stopwatch.Restart();
                 EnableProxyTypes();
+                timers[nameof(EnableProxyTypes)] = stopwatch.ElapsedMilliseconds;
             }
-        }
 
+            stopwatch.Stop();
+        }
 
         /// <summary>
         /// Enable early-bound types from the given context